ecommerce_simple/odoo-mercadolibre/requirements.txt
openpyxl
requests
XlsxWriter
python-dateutil
prophet
pandas
numpy
matplotlib
statsmodels
plotly
openai
iso4217
python-amazon-sp-api
python-dotenv
boto3
requests
urllib3
pytz
numpy==1.23.5
PyPDF2==2.12.1
xmltodict